backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
redis
python-dotenv
redis>=4.0.0
